% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_plot.R
\name{create_waterfall_plot}
\alias{create_waterfall_plot}
\title{Function Title}
\usage{
create_waterfall_plot(
  data,
  vars,
  cohort_col = "cohort",
  second_legend = "response",
  cohort_colors = c("cadetblue", "steelblue", "#9E58CC", "red"),
  second_legend_symbols = NULL,
  main_title = "Waterfall plot for Target Lesion",
  ylab_label = "Change from baseline (\%)",
  y_limits = c(-100, 100),
  hline_values = c(20, -30),
  hline_col = "black",
  hline_lwd = 2,
  hline_lty = 3,
  bar_space = 0.5,
  show_date = TRUE,
  date_format = "\%Y-\%m-\%d",
  date_cex = 1,
  date_font = 3,
  legend_cohorts = NULL,
  legend_cohorts_labels = NULL,
  second_legend_title = "Response",
  second_legend_position = "bottomleft",
  second_legend_inset = c(0.02, -0.025),
  second_legend_cex = 0.9,
  legend_cohorts_cex = 0.9,
  legend_cohorts_position = "topright",
  legend_cohorts_inset = c(0.02, 0.02),
  cex_main = 1.2,
  cex_lab = 1.2,
  cex_axis = 1.5,
  symbol_y_offset = 3,
  add_group_separator = TRUE,
  separator_col = "gray50",
  separator_lwd = 2,
  separator_lty = 2
)
}
\description{
Function Description
}
