% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_two_level_comparision.R
\name{perform_pairwise_tests_simple}
\alias{perform_pairwise_tests_simple}
\title{Function Title}
\usage{
perform_pairwise_tests_simple(
  df_expr,
  df_meta,
  test.type = "t.test",
  adjust.p = FALSE,
  method = "BH",
  var.equal = TRUE
)
}
\description{
Function Description
}
