% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_line_plot.R
\name{plot_patient_lines}
\alias{plot_patient_lines}
\title{Function Title}
\usage{
plot_patient_lines(
  data,
  id_col = "ID",
  cohort_col = "cohort",
  days_col = "days",
  size_col = "size",
  calculate_percentage = TRUE,
  percentage_col_name = "percentage",
  reference_day = 0,
  title = "Patient Follow-up Line Plot",
  subtitle = NULL,
  show_author_date = TRUE,
  author_name = "By LHF",
  xlab = "Follow-up Time (days)",
  ylab = "Percentage from Baseline (\%)",
  legend_title = "Cohort",
  palette = "lancet",
  theme_style = "lancet",
  line_alpha = 0.8,
  line_size = 1,
  point_size = 2,
  show_points = TRUE,
  show_baseline_reference = TRUE,
  baseline_reference_y = 100,
  baseline_reference_color = "gray50",
  baseline_reference_type = "dashed",
  show_PD_PR_reference = TRUE,
  PD_threshold = 20,
  PR_threshold = -30,
  PD_reference_color = "red",
  PR_reference_color = "green",
  PD_PR_reference_type = "dashed",
  custom_legend_labels = NULL,
  font_family = "sans",
  font_size_base = 12,
  font_size_title = 14,
  font_size_subtitle = 11,
  font_size_axis = 10,
  font_size_legend = 10
)
}
\description{
Function Description
}
