% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_two_level_comparision.R
\name{visualize_top_differential_features}
\alias{visualize_top_differential_features}
\title{Function Title}
\usage{
visualize_top_differential_features(
  results_df,
  df_expr,
  df_meta = NULL,
  top_n = 10,
  plot_type = c("heatmap", "volcano", "boxplot"),
  specific_comparison = NULL,
  output_file = NULL
)
}
\description{
Function Description
}
